Adds T2 as a Journal tag
Previously we captured anything J* to be a tag for the
journal title. Endnote doesn't follow this all the time
and uses T2 as the journal tag sometimes.

This also fixes a bug where any line (even the data lines)
that started with a J could be misconstrued as the journal
name.